WIP Day 8
Part 1 is done, but part 2 is incorrect.

Part 2 approach is maybe not the best one. I went for recursion, with a Display data class to represent the state. Basically just implemented a bunch of logic checks and keep going over them trying to build a complete picture.

There's something wrong with the implementation, because 2, 3, and 5 seem to get mixed up, and I think the 9 is missing in the output somehow.

Anyway, AoC is over for this year, so I probably won't come back to this until next December when I remember it exists. :)

Enjoy 2022!